reorganized user apis into user, users, and user_session
