+++ app.js
@@ -186,12 +186,12 @@ app.get(config.auth0.callback_uri, controllers.auth0_sign_in_callback.get)
 
 app.options('/api/v1/users', cors())
 app.post('/api/v1/users', cors(), resources.v1.users.post)
-app.get('/api/v1/users', cors(), resources.v1.users.getUsers)
+app.get('/api/v1/users', cors(), resources.v1.users.get)
 app.options('/api/v1/users/:user_id', cors()) // Enable pre-flight request for authorized PUT request
-app.get('/api/v1/users/:user_id', cors(), resources.v1.users.get)
-app.put('/api/v1/users/:user_id', cors(), resources.v1.users.put)
+app.get('/api/v1/users/:user_id', cors(), resources.v1.user.get)
+app.put('/api/v1/users/:user_id', cors(), resources.v1.user.put)
 app.options('/api/v1/users/:user_id/login-token', cors())
-app.delete('/api/v1/users/:user_id/login-token', cors(), resources.v1.users.delete)
+app.delete('/api/v1/users/:user_id/login-token', cors(), resources.v1.user_session.delete)
 app.get('/api/v1/users/:user_id/streets', cors(), resources.v1.users_streets.get)
 
 app.post('/api/v1/streets', resources.v1.streets.post)
